+++ src/rosdep2/shell_utils.py
@@ -84,7 +84,8 @@ def create_tempfile_from_string_and_execute(string_script, path=None, exec_fn=No
 
     result = 1
     try:
-        fh = tempfile.NamedTemporaryFile('w', delete=False)
+        script_ext = '.bat' if os.name == 'nt' else ''
+        fh = tempfile.NamedTemporaryFile('w', suffix=script_ext, delete=False)
         fh.write(string_script)
         fh.close()
         rd_debug('Executing script below with cwd=%s\n{{{\n%s\n}}}\n' % (path, string_script))
@@ -95,7 +96,7 @@ def create_tempfile_from_string_and_execute(string_script, path=None, exec_fn=No
             else:
                 result = exec_fn(fh.name, cwd=path)
         except OSError as ex:
-            print('Execution failed with OSError: %s' % (ex))
+            print('Execution failed with OSError: %s' % (ex), file=sys.stderr)
     finally:
         if os.path.exists(fh.name):
             os.remove(fh.name)