+++ test/test_rosdep_gentoo.py
@@ -29,7 +29,10 @@
 
 import os
 import traceback
-from mock import Mock, patch
+try:
+    from unittest.mock import Mock, patch
+except ImportError:
+    from mock import Mock, patch
 
 import rospkg.os_detect