+++ test/test_rosdep_lookup.py
@@ -189,7 +189,7 @@ def test_RosdepDefinition():
     except ResolutionError as e:
         assert e.rosdep_key == 'trusty_only_key'
         assert e.os_name == 'ubuntu'
-        assert e.os_version == '*'
+        assert e.os_version == 'lucid'
         # tripwire
         str(e)
     try:
@@ -201,6 +201,41 @@ def test_RosdepDefinition():
         assert e.os_version == 'stretch'
         # tripwire
         str(e)
+
+    definition = RosdepDefinition('invalid_os_wildcard', {'*': ['pytest']}, 'os_wildcard.txt')
+    try:
+        val = definition.get_rule_for_platform('debian', 'sid', ['apt', 'source', 'pip'], 'apt')
+        assert False, 'should have raised: %s' % (str(val))
+    except InvalidData:
+        pass
+
+    definition = RosdepDefinition('non_debian_key', {'debian': None, '*': {'pip': ['pytest']}}, 'os_wildcard.txt')
+    try:
+        val = definition.get_rule_for_platform('debian', 'sid', ['apt', 'source', 'pip'], 'apt')
+        assert False, 'should have raised: %s' % (str(val))
+    except ResolutionError as e:
+        assert e.rosdep_key == 'non_debian_key'
+        assert e.os_name == 'debian'
+        assert e.os_version == 'sid'
+        # tripwire
+        str(e)
+
+    # package manager not supported
+    try:
+        val = definition.get_rule_for_platform('ubuntu', 'precise', ['apt', 'source'], 'apt')
+        assert False, 'should have raised: %s' % (str(val))
+    except ResolutionError as e:
+        assert e.rosdep_key == 'non_debian_key'
+        assert e.os_name == 'ubuntu'
+        assert e.os_version == 'precise'
+        # tripwire
+        str(e)
+
+    val = definition.get_rule_for_platform('ubuntu', 'precise', ['apt', 'source', 'pip'], 'apt')
+    assert val == ('pip', ['pytest']), val
+
+    val = definition.get_rule_for_platform('fedora', '35', ['dnf', 'source', 'pip'], 'dnf')
+    assert val == ('pip', ['pytest']), val
 
     # test reverse merging OS things (first is default)
     definition = RosdepDefinition('test', {'debian': 'libtest-dev'}, 'fake-1.txt')
@@ -327,6 +362,47 @@ def test_RosdepLookup_get_rosdeps():
     print(lookup.get_rosdeps('metapackage_with_deps'))
     assert set(lookup.get_rosdeps('metapackage_with_deps')) == set(['catkin', 'simple_catkin_package', 'another_catkin_package'])  # implicit deps won't get included
     assert set(lookup.get_rosdeps('metapackage_with_deps', implicit=False)) == set(['catkin', 'simple_catkin_package', 'another_catkin_package'])
+
+
+def test_RosdepLookup_dependency_types():
+    from rosdep2.loader import RosdepLoader
+    from rosdep2.lookup import RosdepLookup
+    rospack, rosstack = get_test_rospkgs()
+
+    sources_loader = create_test_SourcesListLoader()
+    default_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                     sources_loader=sources_loader)
+    buildtool_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                       sources_loader=sources_loader, dependency_types=['buildtool'])
+    build_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                   sources_loader=sources_loader, dependency_types=['build'])
+    build_export_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                          sources_loader=sources_loader, dependency_types=['build_export'])
+    exec_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                  sources_loader=sources_loader, dependency_types=['exec'])
+    test_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                  sources_loader=sources_loader, dependency_types=['test'])
+    doc_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                 sources_loader=sources_loader, dependency_types=['doc'])
+    mix_lookup = RosdepLookup.create_from_rospkg(rospack=rospack, rosstack=rosstack,
+                                                 sources_loader=sources_loader, dependency_types=['build', 'build_export'])
+
+    buildtool_deps = ['catkin']
+    build_deps = ['testboost', 'eigen']
+    build_export_deps = ['eigen', 'testtinyxml']
+    exec_deps = ['eigen', 'testlibtool']
+    test_deps = ['curl']
+    doc_deps = ['epydoc']
+    default_deps = buildtool_deps + build_deps + build_export_deps + exec_deps + test_deps
+
+    assert set(buildtool_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(buildtool_deps)
+    assert set(build_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(build_deps)
+    assert set(build_export_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(build_export_deps)
+    assert set(exec_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(exec_deps)
+    assert set(test_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(test_deps)
+    assert set(mix_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(build_deps + build_export_deps)
+    assert set(default_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(default_deps)
+    assert set(doc_lookup.get_rosdeps('multi_dep_type_catkin_package')) == set(doc_deps)
 
 
 def test_RosdepLookup_get_resources_that_need():