+++ test/test_rosdep_main.py
@@ -38,8 +38,10 @@ import rospkg.os_detect
 
 import unittest
 
-from mock import patch
-from mock import DEFAULT
+try:
+    from unittest.mock import DEFAULT, patch
+except ImportError:
+    from mock import DEFAULT, patch
 
 from rosdep2 import main
 from rosdep2.ament_packages import AMENT_PREFIX_PATH_ENV_VAR
@@ -172,8 +174,8 @@ class TestRosdepMain(unittest.TestCase):
             if cmd[0] == 'apt-cache' and cmd[1] == 'showpkg':
                 result = ''
             elif cmd[0] == 'dpkg-query':
-                if cmd[-1] == 'python-dev':
-                    result = '\'python-dev install ok installed\n\''
+                if cmd[-1] == 'python3-dev':
+                    result = '\'python3-dev install ok installed\n\''
                 else:
                     result = '\n'.join(["dpkg-query: no packages found matching %s" % f for f in cmd[3:]])
 
@@ -205,7 +207,12 @@ class TestRosdepMain(unittest.TestCase):
                 expected = [
                     '#[apt] Installation commands:',
                     '  sudo -H apt-get install ros-fuerte-catkin',
-                    '  sudo -H apt-get install libboost1.40-all-dev'
+                    '  sudo -H apt-get install libboost1.40-all-dev',
+                    '  sudo -H apt-get install libeigen3-dev',
+                    '  sudo -H apt-get install libtinyxml-dev',
+                    '  sudo -H apt-get install libltdl-dev',
+                    '  sudo -H apt-get install libtool',
+                    '  sudo -H apt-get install libcurl4-openssl-dev',
                 ]
                 lines = stdout.getvalue().splitlines()
                 assert set(lines) == set(expected), lines
@@ -269,6 +276,12 @@ class TestRosdepMain(unittest.TestCase):
                 rosdep_main(['keys', 'another_catkin_package'] + cmd_extras + ['-i'])
                 stdout, stderr = b
                 assert stdout.getvalue().strip() == 'catkin', stdout.getvalue()
+            with fakeout() as b:
+                rosdep_main(['keys', 'multi_dep_type_catkin_package', '-t', 'test', '-t', 'doc'] + cmd_extras)
+                stdout, stderr = b
+                output_keys = set(stdout.getvalue().split())
+                expected_keys = set(['curl', 'epydoc'])
+                assert output_keys == expected_keys, stdout.getvalue()
         except SystemExit:
             assert False, 'system exit occurred'
         try:
@@ -296,7 +309,7 @@ class TestRosdepMain(unittest.TestCase):
                 rosdep_main(['install', '--from-path', test_package_dir])
                 exit_mock.assert_called_with(1)
             stdout, stderr = b
-            output = stdout.getvalue().splitlines()
-            assert len(output) == 2
-            assert test_package_dir in output[0]
-            assert 'Package version ":{version}" does not follow version conventions' in output[1]
+            output = stderr.getvalue().splitlines()
+            assert len(output) >= 2
+            assert test_package_dir in output[-2]
+            assert 'Package version ":{version}" does not follow version conventions' in output[-1]